Reorganized all the commands
